requirements.txt
Django==3.0.7
psycopg2-binary==2.8.5
django-pagedown==2.0.3
mistune==2.0.0a4
pygments==2.6.1
django-crispy-forms==1.9.1
gunicorn==20.0.4
dj-database-url==0.5.0
whitenoise==5.0.1
django_taggit==1.3.0
whoosh==2.7.4
django-haystack==3.0b2
django-recaptcha==2.0.6
